refactor(prefs): 提取搜索业务类型到 search 子模块
